The Description of the App
iRemind is an application that enables users to manage and organize their events in day-to-day life creatively.
It is straightforward and intuitive to use the app and create To-Do Tasks with reminders. 
It can be wisely used as notes instead of reminders; by deactivating the alarm to keep the event as a note.
It uses Google Maps to make it accessible for the users to select addresses for the locations of the tasks. 
Furthermore, it provides flexibility and effectiveness by allowing users to get directions from their current 
location to the site of the event.